Move utils files into utils submodule
