+++ summarizer/create_split.py
@@ -1,14 +1,3 @@
-''''
-Courtesy of KaiyangZhou
-https://github.com/KaiyangZhou/pytorch-vsumm-reinforce
-
-@article{zhou2017reinforcevsumm,
-   title={Deep Reinforcement Learning for Unsupervised Video Summarization with Diversity-Representativeness Reward},
-   author={Zhou, Kaiyang and Qiao, Yu and Xiang, Tao},
-   journal={arXiv:1801.00054},
-   year={2017}
-}
-'''
 
 from __future__ import print_function
 import os
@@ -20,15 +9,10 @@ import numpy as np
 import errno
 import json
 
-
-parser = argparse.ArgumentParser("Code to create splits in json form")
-parser.add_argument('-d', '--dataset', type=str, required=True, help="path to h5 dataset (required)")
-parser.add_argument('--save-dir', type=str, default='datasets', help="path to save output json file (default: 'datasets/')")
-parser.add_argument('--save-name', type=str, default='splits', help="name to save as, excluding extension (default: 'splits')")
-parser.add_argument('--num-splits', type=int, default=5, help="how many splits to generate (default: 5)")
-parser.add_argument('--train-percent', type=float, default=0.8, help="percentage of training data (default: 0.8)")
-
-args = parser.parse_args()
+"""
+Courtesy of KaiyangZhou
+https://github.com/KaiyangZhou/pytorch-vsumm-reinforce
+"""
 
 def mkdir_if_missing(directory):
     if not osp.exists(directory):
@@ -40,9 +24,8 @@ def mkdir_if_missing(directory):
 
 def write_json(obj, fpath):
     mkdir_if_missing(osp.dirname(fpath))
-    with open(fpath, 'w') as f:
-        json.dump(obj, f, indent=4, separators=(',', ': '))
-
+    with open(fpath, "w") as f:
+        json.dump(obj, f, indent=4, separators=(",", ": "))
 
 def split_random(keys, num_videos, num_train):
     """Random split"""
@@ -55,35 +38,39 @@ def split_random(keys, num_videos, num_train):
             test_keys.append(key)
 
     assert len(set(train_keys) & set(test_keys)) == 0, "Error: train_keys and test_keys overlap"
-
     return train_keys, test_keys
 
-def create():
-    print("==========\nArgs:{}\n==========".format(args))
-    print("Goal: randomly split data for {} times, {:.1%} for training and the rest for testing".format(args.num_splits, args.train_percent))
-    print("Loading dataset from {}".format(args.dataset))
-    dataset = h5py.File(args.dataset, 'r')
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser("Code to create splits in json form")
+    parser.add_argument("-d", "--dataset", type=str, required=True, help="path to h5 dataset (required)")
+    parser.add_argument("--save-dir", type=str, default="splits", help="path to save output json file (default: 'datasets/')")
+    parser.add_argument("--save-name", type=str, default="new_split", help="name to save as, excluding extension (default: 'splits')")
+    parser.add_argument("--num-splits", type=int, default=5, help="how many splits to generate (default: 5)")
+    parser.add_argument("--train-percent", type=float, default=0.8, help="percentage of training data (default: 0.8)")
+    args = parser.parse_args()
+
+    print(f"==========\nArgs:{args}\n==========")
+    print(f"Goal: randomly split data for {args.num_splits} times, {args.train_percent:.1%} for training and the rest for testing")
+    print(f"Loading dataset from {args.dataset}")
+    dataset = h5py.File(args.dataset, "r")
     keys = dataset.keys()
     num_videos = len(keys)
     num_train = int(math.ceil(num_videos * args.train_percent))
     num_test = num_videos - num_train
 
-    print("Split breakdown: # total videos {}. # train videos {}. # test videos {}".format(num_videos, num_train, num_test))
+    print(f"Split breakdown: # total videos {num_videos}. # train videos {num_train}. # test videos {num_test}")
     splits = []
 
     for _ in range(args.num_splits):
         train_keys, test_keys = split_random(keys, num_videos, num_train)
         splits.append({
-            'train_keys': train_keys,
-            'test_keys': test_keys,
+            "train_keys": train_keys,
+            "test_keys": test_keys,
             })
 
-    saveto = osp.join(args.save_dir, args.save_name + '.json')
+    saveto = osp.join(args.save_dir, f"{args.save_name}.json")
     write_json(splits, saveto)
-    print("Splits saved to {}".format(saveto))
+    print(f"Splits saved to {saveto}")
 
     dataset.close()
-
-if __name__ == '__main__':
-    np.random.seed(1994)
-    create()
