+++ summarizer/models/__init__.py
@@ -4,7 +4,7 @@ import numpy as np
 import h5py
 import torch
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from summarizer.vsum_tools import generate_summary, evaluate_summary, generate_scores, evaluate_scores
+from summarizer.utils.eval import generate_summary, evaluate_summary, generate_scores, evaluate_scores
 
 class Trainer:
     """Abstract class handling the training process"""
