+++ summarizer/utils/knapsack.py
@@ -1,0 +1,24 @@
+import numpy as np
+from ortools.algorithms import pywrapknapsack_solver
+
+
+def knapsack_ortools(values, weights, items, capacity):
+    """0-1 Knapsack problem solver"""
+    osolver = pywrapknapsack_solver.KnapsackSolver(
+        pywrapknapsack_solver.KnapsackSolver.KNAPSACK_DYNAMIC_PROGRAMMING_SOLVER,
+        "SummarizationSegmentSelection")
+
+    scale = 1000
+    values = np.array(values)
+    weights = np.array(weights)
+    values = (values * scale).astype(np.int)
+    weights = (weights).astype(np.int)
+    capacity = capacity
+
+    osolver.Init(values.tolist(), [weights.tolist()], [capacity])
+    osolver.Solve()
+    packed_items = [x for x in range(0, len(weights))
+                    if osolver.BestSolutionContains(x)]
+
+    return packed_items
+
